backend/mock-sensor-hub/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
python-multipart
websockets
pyyaml
python-dateutil
numpy
matplotlib
pandas
requests
pyserial
pynmea2
